// frontend/src/app/referral-page/models/referral-state.model.ts
export interface ReferralState {
  createReferral: {
    response: any;
    error: any;
    isLoading: boolean;
  };
  referrals: {
    response: any;
    error: any;
    isLoading: boolean;
  };
  referral: {
    response: any;
    error: any;
    isLoading: boolean;
  };
  userReferral: {
    response: any;
    error: any;
    isLoading: boolean;
  };
}